fix: invalidateblock dogecoin-cli command crashing (#37)
This PR:
- removes calling `getblockcount` after calling `invalidateblock`. An
alternative could be to wait for the daemon to finish invalidating block
cleanly. However, computing the block count is actually not necessary
(users should verify themselves that the chain was fully synced) and the
easy way is to simply to remove it.
- removes backing up the data. Since Dogecoin is not run in prune mode,
the disk space requirement is quite high and we cannot afford having the
data copied and backed up.