fix: utxo-dump tool silently skipping entries (#38)
The file descriptor limit (rlimit) must be set to a high enough limit so
that the utxo-dump tool is able to iterate over all entries of the
chainstate level DB. Otherwise, it may silently fail and skip some
entries. This PR set the rlimit to 16,384, which is, as we observed,
twice the number of *.ldb file for mainnet at height 5,900,000.