fix(bootstrap): correct anchor height parameter (#39)
In the Dogecoin canister, the anchor height corresponds to the height of
the first unstable block in the unstable block tree. During the
bootstrap process, the `HEIGHT` parameter provided by users corresponds
to the height of the stable blocks/UTXO set. The `ANCHOR_HEIGHT`
parameter in `6_compute_canister_state.bin` was set to `HEIGHT - 11`
which is wrong. This PR corrects it to `HEIGHT + 1`.