fix: use buffered writer during pre-upgrade and set stability threshold to 360 to reduce heap memory pressure (#41)
Previously, the canister would serialize the state into an array before writing it to stable memory during pre-upgrade. This approach was inefficient as it constrained the canister to only 2GB in heap memory to make it upgradable. This PR fixes this by using a BufferedWriter to serialize the state in stable memory. The size of the buffer is set to 8MiB.

Additionally, this PR sets the default stability threshold value of the canister to 360, i.e. 6 hours of blocks (compared to 1,440 before). This has the effect of reducing the memory pressure on the heap by ~4. Our experiments showed that this is necessary to prevent the heap from growing beyond 4GB.