fix: backward-compatible state deserialization (#44)
This PR fixes an issue with the deserialization of the state during
`post_upgrade`. The serialization format of the state was modified in
the following PR https://github.com/dfinity/dogecoin-canister/pull/41.
Because the [pre-computed
state](https://dashboard.internetcomputer.org/proposal/138938) has been
serialized using the old format, deserialization now fails. To fix this,
`post_upgrade` now tries to first deserialize the state using the new
format (with `BufferedReader`) and if it fails, tries to deserialize
using the old format.