chore: add proposal `dogecoin_canister_upgrade_2025_10_24.md` (#46)
Proposal to upgrade the Dogecoin canister for the launch of the Dogecoin
canister.

The [previous
proposal](https://dashboard.internetcomputer.org/proposal/138995) to
install the Dogecoin canister was accepted but due to a bug related to
the deserialization of the canister state (see
https://github.com/dfinity/dogecoin-canister/pull/44), the installation
of the WASM failed.

This proposal contains the fix to ensure correct deserialization of the
state that was uploaded using the `uploader` canister. After the
Dogecoin canister starts, it will sync blocks from height 5,906,189 up
to the tip, after which its API will become accessible.