docs: skeletton for developer docs (#47)
Developer documentation for how to build on Dogecoin from ICP will be
hosted in this repository. This PR starts the work on the developer
documentation by:

1. Adding the developer documentation under `/docs`. Right now it only
contains the section titles, following the organization on [Build on
Bitcoin](https://internetcomputer.org/docs/build-on-btc/).
2. Using [mdBook](https://rust-lang.github.io/mdBook/), since the docs
needs to contain code samples from other files, corresponding to
examples (similar to
[basic_bitcoin](https://github.com/dfinity/examples/tree/master/rust/basic_bitcoin)).
The setup is similar to the one used for [stable
structures](https://dfinity.github.io/stable-structures/).
3. Adding a CI job to publish the book to the Github page:
https://dfinity.github.io/dogecoin-canister