feat: add state-reader tool (#29)
Adds a state-reader tool that can be used to read the Dogecoin canister
snapshot. The tool is able to parse all stable memory regions, compute
their hash, and prints useful information on the state (such as UTXO
value distribution, address balance distribution, etc.). It is useful to
verify the consistency between the state obtained when letting the
canister sync and the one pre-computed using the state-builder tool (in
particular the [utxo-dump
tool](https://github.com/dfinity/dogecoin-canister/pull/24)).