feat: use 1,000 transactions in the percentiles calculation (#60)
From @THLO: "The `dogecoin_get_current_fee_percentiles` returns the
percentiles of fee rates just like for Bitcoin. However, there are
typically significantly fewer transactions per block, about 25
transactions per block on average. For Bitcoin, the most recent 10,000
transactions are considered for the fee percentiles, i.e., the
transactions in the past approximately 40 minutes as each block contains
about 2500 transactions.
Dogecoin has a block rate that is 10 times higher at 1 block every
minute on average but the blocks are smaller by 2 orders of magnitude.
As a result, in order to consider transactions over a similar timespan
as for Bitcoin, the most recent 1,000 transactions must be considered
for the computation of the fee percentiles."

[DEFI-2528](https://dfinity.atlassian.net/browse/DEFI-2528): This PR
reduces the number of transactions considered in the fee percentiles
computation from 10,000 to 1,000.


[DEFI-2528]:
https://dfinity.atlassian.net/browse/DEFI-2528?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ