refactor!: remove testnet (#61)
As testnet is not supported, this PR removes the testnet network variant
from the Dogecoin canister interface. Note that introducing testnet
later on if required would not be a breaking change, except for the
get_config endpoint (added a comment in the .did file that it is a debug
endpoint).

As this PR removes the testnet variant from network which is used in
several requests (e.g. `get_balance_request`), then this PR introduces a
**BREAKING CHANGE**.