chore: Cleaning up the interface specification (#56)
The interface specification was copied over from the Bitcoin canister
interface specification, and "Bitcoin" was replaced with "Dogecoin".

However, there are a few differences, for example in terms of
parameterization, between the Bitcoin and Dogecoin canisters. This PR
updates the interface specification, fixing inaccuracies due to the
differences between the canisters.

This PR addresses ticket
[DEFI-2025](https://dfinity.atlassian.net/browse/DEFI-2025).

[DEFI-2025]:
https://dfinity.atlassian.net/browse/DEFI-2025?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ