chore: proposal to upgrade the dogecoin canister to release/2025-12-10 (#65)
This proposal aims to upgrade the canister to [release
2025-12-10](https://github.com/dfinity/dogecoin-canister/releases/tag/release%2F2025-12-10).

## Motivation

This proposal contains several upstream cherry-picks to upgrade the
`ic-cdk` version used to `v0.19.0` and add a tool to verify the
endpoints the canister’s WASM exports against its Candid interface.

Additionally, it changes the number of transactions used in the fee
percentiles calculation from 10,000 to 1,000, to make sure the same
timespan of transaction is considered in the calculation as the Bitcoin
canister (approximately 40 minutes).

This proposal also removes testnet from the interface of the canister as
the Dogecoin testnet network is not supported.

Finally, it raises the stability threshold from 720 to 1,440,
corresponding to an average of one day worth of blocks, for additional
security.