feat: add network validation for addresses in get_balance and get_utxos requests (#458)
[DEFI-2519](https://dfinity.atlassian.net/browse/DEFI-2519?atlOrigin=eyJpIjoiNDc5ODg4MDk2YWViNGE1YWE3MGIxZWYxOTRlMGVjNjMiLCJwIjoiaiJ9):
When `bitcoin_get_balance` and `bitcoin_get_utxos` are called and the
address does not match a known format (e.g. P2PKH, etc.), the canister
returns a `MalformedAddress` error. However, no such error is returned
when the address is an address for the wrong network (e.g. testnet
instead of mainnet). This can be confusing for users trying to get the
balance of e.g. a testnet address from the mainnet canister and getting
0 as balance. This PR adds the `from_str_checked` which is used in
`get_balance` and `get_utxos` to validate that the address in the
request belongs to the same network as the canister called.

[DEFI-2519]:
https://dfinity.atlassian.net/browse/DEFI-2519?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ